user_form
This project was bootstrapped with Create React App.

Create a Django or Node.js or Go or Ruby on Rails or Spring Application which shows a simple form at "/user-form". The form asks the user for his details(name, date of birth, email, phone number). The form is rendered in any of the modern frontend frameworks like AngularJs, React.js or Vue.js. Date of birth field should be a calendar option.

User fills the form and presses submit. Perform basic front-end validation on the name, email, DoB (age cannot be less than 18 years).

Phone number validation happens at the backend. If you are using Django then Django-Rest-Framework is preferred. If node.js or Go or Scala, use the relevant framework/suggested best practice for writing a modular code.